Create escjail
CAUTION:::HINTS BELOW 
There are two cases 
case 1: we reach a door but we dont have the key . in this case we cannot push this node in queue. we visit the key required later for this node . That is the time when we push this particular door in queue. 
case 2: we reach a door and we already reached the key . we simply push the door in queue. 
Dont try this with dfs